move button to own dir. update resize script.
